Add osc-placement-dsvm-functional-py3 job
This adds a py3 variant of the functional dsvm job.
This still uses devstack-gate (not native zuulv3) because
I'm not yet cool enough to know how to do native zuulv3 things.

Swift has to be disabled since it doesn't support py3 (this is
something all other py3 jobs are doing right now too).

Change-Id: I9a0074d9658eab76b7c73b820f498d61b89545e7