Provide a useful message in case of 5xx error
If server response does not have an appropriate json format,
the message was useless. For example, if placement-api is down,
we had something like this:

$ openstack resource provider delete
Expecting value: line 1 column 1 (char 0)

This was because the code assumed that the response had details in
it. That is correct for 4xx errors, but that's not the case for 5xx
errors. This patch changes it not to retrieve the details from
response on 5xx error, and just raise keystone exceptions.

Now we have something like this:

$ openstack resource provider delete
Service Unavailable (HTTP 503)

Story: #2005437
Story: #2005438
Task: #30478
Task: #30479

Change-Id: I07fa0b626b7529699179c8cd8e254fd66455c930
Co-Autored-By: Tetsuro Nakamura <tetsuro.nakamura.bc@hco.ntt.co.jp>