Add support for 1.18 microversion
The 1.18 microversion of placement adds support for the
required parameter when requesting a list of resource
providers. The values are traits which limit the results
to only those providers that have those traits. The
--required argument may be specified multiple times.

Note the NOTE in the added test about raciness. There's
not a ton we can do about that since the functional
tests aren't actually functional tests, they are
integration.

Change-Id: If79b4d12c712fdeeafa5ea77ae3411d31c8cfe9b