Add support for 1.19 microversion
The 1.19 microversion of placement adds support for the
generation parameter when setting aggregates on a resource
provider. The value is an integer generation, available
when listing or showing resource providers.

If an earlier version is chosen, but a --generation arg
provided, an error is reported. Similarly, if not generation
is provided and the version is 1.19 or later, an error
is reported.

The new TestAggregate119 inherits from TestAggregate to
repeat the general validation it does, overrides those
methods which require a generation when using version
1.19, and adds two failure tests. One test in the super
class is adjusted so that it works as desired in both
classes and in both versions of Python (Python3 has
different behavor of a *args before kw args). I considered
changing the signature of base.resource_provider_aggregate_set
but that would have meant changes is several places.

Change-Id: I74157998dd19d8a8d2b68d12ed3726aac11eb013