Microversion 1.21 support
Add support for member_of/aggregate-uuid when listing allocation
candidates. The choice of arguments intentionally mirrors those
used when listing resource providers, which means that the term
'member of' is not used.

Note that microversion 1.24 [1] is going to make argument parsing
in this area considerably more complex. Not for the computer, but
the humans using the tool. For now we probably want to do what's
here, to do the aforementioned mirroring, but when we get to 1.24
we'll need some brains.

[1] https://docs.openstack.org/placement/latest/placement-api-microversion-history.html#support-multiple-member-of-queryparams

Change-Id: Iba3e31fec29e8e85aa2f6a4248f0a27d7373ba6d