Only enable keystone and placement for functional test runs
The functional test jobs require a devstack environment which
makes the job longer, but we should really only need keystone
and placement (and a database) so this change explicitly sets
the ENABLED_SERVICES list that will be used in devstack.

As a result, the task to disable swift from the py3 job can
be removed.

Change-Id: Ifbd9c1777b2d565f95e42f6264e2e10fd1679f33
Story: #2005411