Build pdf docs
This follows the instructions [1] in an attempt to build pdf docs.

To get a build, the "badge" in the README has been removed.
The other option to add some svg -> pdf support, but that seems
overkill for a badge which is effectively meaningless.

The build wanted 'xindy' to build an index, but we can turn that
off and still get useful results.

[1] https://etherpad.openstack.org/p/train-pdf-support-goal

Change-Id: Icf7c22bf9d1de6fb2a74a756c370930d4c00b0b9
Story: 2006110
Task: 35397