Use os-endpoint instead of os-url for functional tests
The functional tests use an explicit endpoint and token for auth
to the placement fixture. Before openstack client 4.0.0 os-url
was a way to specify the endpoint. Now os-endpoint is the way.

Update accordingly.

Note that this doesn't impact the behavior of the plugin itself,
just the tests.

Change-Id: Id4f8f0224f61c1741e6a1a8369ec053bb1824e8a
(cherry picked from commit f49b9fcc81918ca937ed7332702432c3f82eb639)