Be explicit about auth type in functional tests
When using the old style of '--os-url' we didn't need to be
explicit, it deduced the right type. And python-openstackclient
>= 4.0.0 it does for --os-endpoint. But when using '--os-endpoint'
with <4.0.0 we need to state 'admin_token' explicitly.

This form works in both 3.19.x and 4.0.x.

Change-Id: Ie9c76a23658b46997dcb17ba3fb7f1e700eca506
(cherry picked from commit 306a91e9e8939d75e0309e492502fa0805cc2c8d)