Add resource provider allocation unset command
Adds the "openstack resource provider allocation unset" command which
allows removing resource provider(s) from the set of allocations for
the given consumer. This can be useful when a consumer has allocations
against more than one resource provider and
"openstack resource provider allocation delete" is undesirable as it
removes all allocations for the consumer. The new unset command requires
"--os-placement-api-version 1.12" or greater.

Note that the --provider option can be specified multiple times.

If the resulting allocations are all gone and 1.28 or greater is not
being used, a DELETE is called rather than PUT since
PUT /allocations/{consumer_id} before 1.28 does not allow an empty
"allocations" dict in the request body. In this case nothing is output
similar to the "openstack resource provider allocation delete" command.

Change-Id: I627bfd1ff699d075028da6afafbe7fb9b2f13058
Story: #2006779
Task: #37304