Support auto-negotiated microversion
To improve the usability for those who are not familiar with
microversion, this patch adds microversion negotiation to pick up max
microversion supported both by client and by server.

This negotiation is enabled by specifying only major api version,
``--os-placement-api-version 1``. Note that the default microversion
remains to be ``1.0`` in this patch and will be switched to the
negotiated version in the following patch.

Change-Id: I2998ff0f3941bf226a942969adf75564a8d5a065
Story: #2005448
Task: #30497