Support granular allocation candidate list
The ``openstack allocation candidate list`` command now supports
``--group`` and ``--group-policy`` option. The ``--group`` option is
accepts integer to group granular requests. If specified, following
given options of resources, required/forbidden traits, and aggregates
are associated to that group and will be satisfied by the same resource
provider in the response. Can be repeated to get candidates from
multiple resource providers in the same resource provider tree.

If multiple groups are supplied, the separate groups may or may not
be satisfied by the same provider. If you want the groups to be
satisfied by different providers, set ``--group_policy`` to
``isolate``.

For example:
    ``openstack allocation candidate list \
        --group 1 --resource VCPU=3 --required HW_CPU_X86_SSE \
        --group 2 --resource VCPU=4 \
        --group_policy isolate``

The granular request is supported in 1.25, but this is tested via
microversion 1.29 since the API is unaware of nested provider
before that version. Note that 1.26-1.29 has no impact to client:
  1.25 - Adds support for granular resource requests via numbered
         querystring groups in GET /allocation_candidates.
  1.26 - Add ability to specify inventory with reserved value equal
         to total.
  1.27 - Include all resource class inventories in `provider_summaries`
         field in response of `GET /allocation_candidates` API even if
         the resource class is not in the requested resources.
  1.28 - Add support for consumer generation
  1.29 - Support nested providers in `GET /allocation_candidates` API.

Change-Id: Ib32181abd12db0c2983065ea9846e911033d53f1