Support multiple member_of query parameter
The ``openstack allocation candidate list`` and the ``openstack
allocation candidate list`` command now supports ``--member-of``
option, which accepts a list of comma-separated UUIDs of the
resource provider aggregates. If this is specified, the returned
resource providers must be associated with at least one of the
aggregates identified by uuid. This option can be repeated to
add(restrict) the condition with ``--os-placement-api-version 1.24``
or greater.

This deprecates the ``--aggregate-uuid`` option because it is
difficult to keep using this option to support multiple ``member_of``
feature added in microversion 1.24.

Change-Id: I0f3068c411c504380918f27f994734528fe55c77
Co-Autored-By: Tetsuro Nakamura <tetsuro.nakamura.bc@hco.ntt.co.jp>
Story: #2006940
Task: #37613