Add Python3 zed unit tests
This is an automatically generated patch to ensure unit testing
is in place for all the of the tested runtimes for zed.

See also the PTI in governance [1].

I needed to squash in another change that is dropping the py36 and py37
support. We cannot do that separately as those jobs are already failing.

[1]: https://governance.openstack.org/tc/reference/project-testing-interface.html

Change-Id: I378d5c3c1f3e96696670a4865caf5901269a5ee4