Use pypi released version of placement in functional tests
osc-placement functional tests are using the master version
of placement which is failing due to the constraints mismatch
on stable branches testing.

- https://zuul.opendev.org/t/openstack/build/a2cc830f28f44ed5ae0e1c5f0a99ac75

During the discussion in Nova meeting, we did not find any reason why
we are using the latest placement on stable branch testing of osc-placement
- https://meetings.opendev.org/meetings/nova/2023/nova.2023-01-10-16.00.log.html#l-47

Below is the original patch which started using the placement from master
source.
- https://review.opendev.org/c/openstack/osc-placement/+/651939

To avoid the constraints mismatch on stable branch testing which can
happen in future also, let's use it from pypi version which will be the
same way as nova functional tests doing
- https://github.com/openstack/nova/blob/26f24b7086855bc8f57059d618060cc5844dca15/tox.ini#L64

During this backport, removing basepython changes are not applicable on stable as we have
pinned stable branches for tox<4.

Change-Id: I4e3e5732411639054baaa9211a29e2e2c8210ac0
(cherry picked from commit 90ecfa92f030a6370658430ca3bbc6273f7ea487)