Remove use of distutils
This has been removed in Python 3.12, which is preventing us using
osc-placement in those environments. Thankfully we only use it for
StrictVersion which we are using to parse microversions. This is
actually overkill here since we can always expect a version string like
'1.5', thus we do this parsing ourselves rather than drag in a new
library like 'microversion_parse'.

Change-Id: I2923238f8b30568bceb1799882d4468ff1d7e538
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>