[stable-only] Cap setuptools <71.0.0
py38 jobs (on ubuntu-focal) started to fail (see the relevant github
issue [1]) due to recent virtualenv release (20.26.4; which bundles
setuptools).

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

Also, do not install placement from master branch, rather from
unmaintained/victoria.

[1] https://github.com/pypa/setuptools/issues/4483

Change-Id: I83144e9fb5d2decbe33b382daffe795eb43ee8d7